+++ cognite/client/_api/events.py
@@ -141,22 +141,24 @@ class EventsAPI(APIClient):
         created_time: Dict[str, Any] = None,
         last_updated_time: Dict[str, Any] = None,
         external_id_prefix: str = None,
+        partitions: int = None,
         limit: int = 25,
     ) -> EventList:
         """List events
 
         Args:
-            start_time (Dict[str, Any]): Range between two timestamps
-            end_time (Dict[str, Any]): Range between two timestamps
+            start_time (Dict[str, Any]): Range between two timestamps.
+            end_time (Dict[str, Any]): Range between two timestamps.
             type (str): Type of the event, e.g 'failure'.
             subtype (str): Subtype of the event, e.g 'electrical'.
             metadata (Dict[str, Any]): Customizable extra data about the event. String key -> String value.
             asset_ids (List[int]): Asset IDs of related equipments that this event relates to.
             root_asset_ids (List[Dict[str, Any]]): The IDs of the root assets that the related assets should be children of.
             source (str): The source of this event.
-            created_time (Dict[str, Any]): Range between two timestamps
-            last_updated_time (Dict[str, Any]): Range between two timestamps
-            external_id_prefix (str): External Id provided by client. Should be unique within the project
+            created_time (Dict[str, Any]): Range between two timestamps.
+            last_updated_time (Dict[str, Any]): Range between two timestamps.
+            external_id_prefix (str): External Id provided by client. Should be unique within the project.
+            partitions (int): Retrieve events in parallel using this number of workers. Also requires `limit=None` to be passed.
             limit (int, optional): Maximum number of events to return. Defaults to 25. Set to -1, float("inf") or None
                 to return all items.
 
@@ -198,7 +200,7 @@ class EventsAPI(APIClient):
             type=type,
             subtype=subtype,
         ).dump(camel_case=True)
-        return self._list(method="POST", limit=limit, filter=filter)
+        return self._list(method="POST", limit=limit, filter=filter, partitions=partitions)
 
     def create(self, event: Union[Event, List[Event]]) -> Union[Event, EventList]:
         """Create one or more events.